corrected the limits for the fixation point

+++ src/gazeArbiter/include/iCub/gazeArbiterModule.h
@@ -208,9 +208,9 @@ class gazeArbiterModule:public yarp::os::RFModule {
     int xoffset;                                // offset for the 3D point along x
     int yoffset;                                // offset for the 3D point along y
     int zoffset;                                // offset for the 3D point along z
-    int xmax, xmin;                             // limits for the allowed fixation point (x axis)
-    int ymax, ymin;                             // limits for the allowed fixation point (y axis)
-    int zmax, zmin;                             // limits for the allowed fixation point (z axis)
+    double xmax, xmin;                             // limits for the allowed fixation point (x axis)
+    double ymax, ymin;                             // limits for the allowed fixation point (y axis)
+    double zmax, zmin;                             // limits for the allowed fixation point (z axis)
     int width, height;                          // parameter set by user dimensioning input image
     yarp::os::Port handlerPort;                 // a port to handle messages
